refactor(web): extract MemberListContent/PendingInvitesSection; simplify confirm-dialog copy
Pulls both inline sub-components out of LobbyMemberList into their own
files. Replaces the per-render title/message/confirmLabel/danger ternary
chain with a single MEMBER_ACTION_CONFIG lookup keyed by action kind.
